moved every non tauri stuff to handler
