+++ src-tauri/cabr2_config/src/lib.rs
@@ -3,21 +3,13 @@ mod error;
 mod handler;
 mod types;
 
-use std::{env, path::PathBuf};
+use std::env;
 
-use directories_next::ProjectDirs;
-use lazy_static::lazy_static;
 use tauri::{self, plugin::Plugin};
 
 use cmd::Cmd;
-pub use handler::{read_config, write_config};
+pub use handler::{read_config, write_config, DATA_DIR, PROJECT_DIRS, TMP_DIR};
 pub use types::BackendConfig;
-
-lazy_static! {
-  pub static ref PROJECT_DIRS: ProjectDirs = ProjectDirs::from("de", "Calciumdibromid", "CaBr2").unwrap();
-  pub static ref DATA_DIR: PathBuf = get_program_data_dir();
-  pub static ref TMP_DIR: PathBuf = env::temp_dir();
-}
 
 pub struct Config;
 
@@ -107,55 +99,5 @@ impl Plugin for Config {
 
   fn ready(&self, _: &mut tauri::Webview<'_>) {
     log::trace!("plugin ready");
-  }
-}
-
-fn get_program_data_dir() -> PathBuf {
-  #[cfg(not(debug_assertions))]
-  {
-    #[cfg(not(feature = "portable"))]
-    {
-      #[cfg(target_os = "linux")]
-      {
-        log::trace!("data path: linux");
-        return PathBuf::from("/usr/lib/cabr2/");
-      }
-
-      #[cfg(target_os = "macos")]
-      {
-        log::trace!("data path: macos");
-        unimplemented!();
-      }
-
-      #[cfg(target_os = "windows")]
-      {
-        log::trace!("data path: windows");
-        return PathBuf::from(env::args().next().unwrap())
-          .parent()
-          .unwrap()
-          .to_path_buf();
-      }
-    }
-
-    #[cfg(feature = "portable")]
-    {
-      log::trace!("data path: portable");
-      return PathBuf::from(env::args().next().unwrap())
-        .parent()
-        .unwrap()
-        .to_path_buf();
-    }
-  }
-
-  #[cfg(debug_assertions)]
-  {
-    log::trace!("data path: debug");
-    let mut program_path = PathBuf::from(env::args().next().unwrap())
-      .parent()
-      .unwrap()
-      .to_path_buf();
-    // git repo root
-    program_path.push("../../../");
-    program_path.canonicalize().unwrap()
   }
 }
